Re-organize '/types' directory
