--- src/types/core/index.ts
@@ -1,4 +1,0 @@
-export * from './json-rpc-types';
-export * from './message-types';
-export * from './exception-types';
-export * from './sdk-types';
