+++ src/types/index.ts
@@ -1,2 +1,11 @@
-export * from './core';
-export * from './modules';
+// Only re-export types that are intended for the public API from this file.
+
+export * from './core/json-rpc-types';
+export * from './core/message-types';
+export * from './core/exception-types';
+export * from './core/sdk-types';
+export * from './core/view-types';
+
+export * from './modules/auth-types';
+export * from './modules/user-types';
+export * from './modules/rpc-provider-types';
