--- src/types/internal/index.ts
@@ -1,2 +1,0 @@
-export * from './extension-types';
-export * from './utility-types';
