--- src/types/modules/index.ts
@@ -1,3 +1,0 @@
-export * from './auth-types';
-export * from './user-types';
-export * from './rpc-provider-types';
